Fix lsm recovery (#712)
* Don't continue in face of error when failing to build index

* Don't allow compactions during index close

* index: fix snapshot

Previously we were skipping all index files if the curent index file was
empty. This fixes it to always snapshot the index files and only
conditionally skip snapshotting the active file if emtpy

* remove the index dir on db creation

* lint

+++ db.go
@@ -496,8 +496,13 @@ func (s *ColumnStore) DB(ctx context.Context, name string, opts ...DBOption) (*D
 	}
 
 	if dbSetupErr := func() error {
-		if err := os.RemoveAll(db.trashDir()); err != nil {
-			return err
+		if db.storagePath != "" {
+			if err := os.RemoveAll(db.trashDir()); err != nil {
+				return err
+			}
+			if err := os.RemoveAll(db.indexDir()); err != nil { // Remove the index directory. These are either restored from snapshots or rebuilt from the WAL.
+				return err
+			}
 		}
 		db.txPool = NewTxPool(&db.highWatermark)
 		// Wait to start the compactor pool since benchmarks show that WAL
